ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
pydantic==2.5.2
python-multipart==0.0.6
Pillow>=10.0.0
